InnovaSolutionMVPGestionActivos/requirements.txt
asgiref==3.11.1
dj-rest-auth==7.2.0
Django==6.0.4
django-allauth==65.16.1
django-cors-headers==4.9.0
djangorestframework==3.17.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1